Split stack generation logic
