+++ config/stack.go
@@ -7,6 +7,9 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/gruntwork-io/terragrunt/internal/experiment"
+	"github.com/hashicorp/go-getter/v2"
+
 	"github.com/hashicorp/hcl/v2/hclsyntax"
 
 	"github.com/gruntwork-io/terragrunt/util"
@@ -20,10 +23,12 @@ import (
 )
 
 const (
-	stackDir       = ".terragrunt-stack"
-	unitValuesFile = "terragrunt.values.hcl"
-	unitDirPerm    = 0755
-	valueFilePerm  = 0644
+	stackDir         = ".terragrunt-stack"
+	unitValuesFile   = "terragrunt.values.hcl"
+	manifestName     = ".terragrunt-stack-manifest"
+	defaultStackFile = "terragrunt.stack.hcl"
+	unitDirPerm      = 0755
+	valueFilePerm    = 0644
 )
 
 // StackConfigFile represents the structure of terragrunt.stack.hcl stack file.
@@ -46,6 +51,222 @@ type Stack struct {
 	Name   string `hcl:",label"`
 	Source string `hcl:"source,attr"`
 	Path   string `hcl:"path,attr"`
+}
+
+// GenerateStacks generates the stack files.
+func GenerateStacks(ctx context.Context, opts *options.TerragruntOptions) (error, bool) {
+	processedFiles := make(map[string]bool)
+	// initial files setting as stack file
+	foundFiles := []string{opts.TerragruntStackConfigPath}
+	for {
+		// check if we have already processed the files
+		processedNewFiles := false
+		for _, file := range foundFiles {
+			if processedFiles[file] {
+				continue
+			}
+			processedNewFiles = true
+			processedFiles[file] = true
+			if err := processStackFile(ctx, opts, file); err != nil {
+				return errors.Errorf("Failed to process stack file %s %v", file, err), true
+			}
+		}
+		if !processedNewFiles {
+			break
+		}
+		newFiles, err := listStackFiles(opts, opts.WorkingDir)
+		if err != nil {
+			return errors.Errorf("Failed to list stack files %v", err), true
+		}
+		foundFiles = newFiles
+	}
+	return nil, false
+}
+
+// StackOutput generates the output from the stack files.
+func StackOutput(ctx context.Context, opts *options.TerragruntOptions) (map[string]map[string]cty.Value, error) {
+	opts.Logger.Debugf("Generating output from %s", opts.TerragruntStackConfigPath)
+	opts.TerragruntStackConfigPath = filepath.Join(opts.WorkingDir, defaultStackFile)
+	stackTargetDir := filepath.Join(opts.WorkingDir, stackDir)
+	stackFiles, err := listStackFiles(opts, stackTargetDir)
+	if err != nil {
+		return nil, errors.Errorf("Failed to list stack files in %s %v", stackTargetDir, err)
+	}
+	unitOutputs := make(map[string]map[string]cty.Value)
+
+	// add default stack file
+	stackFiles = append(stackFiles, opts.TerragruntStackConfigPath)
+
+	for _, path := range stackFiles {
+		stackFile, err := ReadStackConfigFile(ctx, opts, path)
+
+		if err != nil {
+			return nil, errors.New(err)
+		}
+
+		// process each unit and get outputs
+		for _, unit := range stackFile.Units {
+			opts.Logger.Debugf("Processing unit %s", unit.Name)
+			dir := filepath.Dir(path)
+			unitDir := filepath.Join(dir, stackDir, unit.Path)
+			output, err := unit.ReadOutputs(ctx, opts, unitDir)
+
+			if err != nil {
+				return nil, errors.New(err)
+			}
+
+			unitOutputs[unit.Name] = output
+		}
+	}
+
+	return unitOutputs, nil
+}
+
+// processStackFile process single stack file.
+func processStackFile(ctx context.Context, opts *options.TerragruntOptions, stackFilePath string) error {
+	stackSourceDir := filepath.Dir(stackFilePath)
+	stackFile, err := ReadStackConfigFile(ctx, opts, stackFilePath)
+	if err != nil {
+		return errors.Errorf("Failed to read stack file %s in %s %v", stackFilePath, stackSourceDir, err)
+	}
+	stackTargetDir := filepath.Join(stackSourceDir, stackDir)
+	if err := os.MkdirAll(stackTargetDir, os.ModePerm); err != nil {
+		return errors.Errorf("failed to create base directory: %s %v", stackTargetDir, err)
+	}
+
+	if err := processUnits(ctx, opts, stackSourceDir, stackTargetDir, stackFile.Units); err != nil {
+		return err
+	}
+
+	if err := processStacks(ctx, opts, stackSourceDir, stackTargetDir, stackFile.Stacks); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// processUnits processes the units in the stack file.
+func processUnits(ctx context.Context, opts *options.TerragruntOptions, stackSourceDir, stackTargetDir string, units []*Unit) error {
+	for _, unit := range units {
+		opts.Logger.Infof("Processing unit %s", unit.Name)
+
+		destPath := filepath.Join(stackTargetDir, unit.Path)
+		dest, err := filepath.Abs(destPath)
+
+		if err != nil {
+			return errors.Errorf("failed to get absolute path for destination '%s': %v", dest, err)
+		}
+
+		src := unit.Source
+		opts.Logger.Debugf("Processing unit: %s (%s) to %s", unit.Name, src, dest)
+
+		if err := copyFiles(ctx, opts, unit.Name, stackSourceDir, src, dest); err != nil {
+			return err
+		}
+
+		// generate unit values file
+		if err := writeUnitValues(opts, unit, dest); err != nil {
+			return errors.Errorf("Failed to write unit values %v %v", unit.Name, err)
+		}
+	}
+
+	return nil
+}
+
+// processStacks processes the stacks in the stack file.
+func processStacks(ctx context.Context, opts *options.TerragruntOptions, stackSourceDir, stackTargetDir string, stacks []*Stack) error {
+	for _, stack := range stacks {
+		opts.Logger.Infof("Processing stack %s", stack.Name)
+
+		destPath := filepath.Join(stackTargetDir, stack.Path)
+		dest, err := filepath.Abs(destPath)
+
+		if err != nil {
+			return errors.Errorf("Failed to get absolute path for destination '%s': %v", dest, err)
+		}
+
+		src := stack.Source
+		opts.Logger.Debugf("Processing stack: %s (%s) to %s", stack.Name, src, dest)
+
+		if err := copyFiles(ctx, opts, stack.Name, stackSourceDir, src, dest); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// copyFiles copies files or directories from a source to a destination path.
+//
+// The function checks if the source is local or remote. If local, it copies the
+// contents of the source directory to the destination. If remote, it fetches the
+// source and stores it in the destination directory.
+func copyFiles(ctx context.Context, opts *options.TerragruntOptions, identifier, sourceDir, src, dest string) error {
+	if isLocal(opts, sourceDir, src) {
+		// check if src is absolute path, if not, join with sourceDir
+		var localSrc string
+		if filepath.IsAbs(src) {
+			localSrc = src
+		} else {
+			localSrc = filepath.Join(sourceDir, src)
+		}
+		localSrc, err := filepath.Abs(localSrc)
+
+		if err != nil {
+			opts.Logger.Warnf("failed to get absolute path for source '%s': %v", identifier, err)
+			// fallback to original source
+			localSrc = src
+		}
+
+		if err := util.CopyFolderContentsWithFilter(opts.Logger, localSrc, dest, manifestName, func(absolutePath string) bool {
+			return true
+		}); err != nil {
+			return errors.Errorf("Failed to copy %s to %s %v", localSrc, dest, err)
+		}
+	} else {
+		if err := os.MkdirAll(dest, os.ModePerm); err != nil {
+			return errors.Errorf("Failed to create directory %s for %s %v", dest, identifier, err)
+		}
+
+		if _, err := getter.GetAny(ctx, dest, src); err != nil {
+			return errors.Errorf("Failed to fetch %s %v", identifier, err)
+		}
+	}
+	return nil
+}
+
+// isLocal determines if a given source path is local or remote.
+//
+// It checks if the provided source file exists locally. If not, it checks if
+// the path is relative to the working directory. If that also fails, the function
+// attempts to detect the source's getter type and recognizes if it is a file URL.
+func isLocal(opts *options.TerragruntOptions, workingDir, src string) bool {
+	// check initially if the source is a local file
+	if util.FileExists(src) {
+		return true
+	}
+
+	src = filepath.Join(workingDir, src)
+	if util.FileExists(src) {
+		return true
+	}
+	// check path through getters
+	req := &getter.Request{
+		Src: src,
+	}
+	for _, g := range getter.Getters {
+		recognized, err := getter.Detect(req, g)
+		if err != nil {
+			opts.Logger.Debugf("Error detecting getter for %s: %v", src, err)
+			continue
+		}
+
+		if recognized {
+			break
+		}
+	}
+
+	return strings.HasPrefix(req.Src, "file://")
 }
 
 // ReadOutputs reads the outputs from the unit.
@@ -102,7 +323,7 @@ func ReadStackConfigFile(ctx context.Context, opts *options.TerragruntOptions, f
 	return config, nil
 }
 
-// WriteUnitValues generates and writes unit values to a terragrunt.values.hcl file in the specified unit directory.
+// writeUnitValues generates and writes unit values to a terragrunt.values.hcl file in the specified unit directory.
 // If the unit has no values (Values is nil), the function logs a debug message and returns.
 // Parameters:
 //   - opts: TerragruntOptions containing logger and other configuration
@@ -110,9 +331,9 @@ func ReadStackConfigFile(ctx context.Context, opts *options.TerragruntOptions, f
 //   - unitDirectory: Target directory where the values file will be created
 //
 // Returns an error if the directory creation or file writing fails.
-func WriteUnitValues(opts *options.TerragruntOptions, unit *Unit, unitDirectory string) error {
+func writeUnitValues(opts *options.TerragruntOptions, unit *Unit, unitDirectory string) error {
 	if unitDirectory == "" {
-		return errors.New("WriteUnitValues: unit directory path cannot be empty")
+		return errors.New("writeUnitValues: unit directory path cannot be empty")
 	}
 
 	if err := os.MkdirAll(unitDirectory, unitDirPerm); err != nil {
@@ -184,145 +405,7 @@ func ReadUnitValues(ctx context.Context, opts *options.TerragruntOptions, unitDi
 	return &result, nil
 }
 
-// ValidateStackConfig validates a StackConfigFile instance according to the rules:
-// - Unit name, source, and path shouldn't be empty
-// - Unit names should be unique
-// - Units shouldn't have duplicate paths
-// - Stack name, source, and path shouldn't be empty
-// - Stack names should be unique
-// - Stack shouldn't have duplicate paths
-func ValidateStackConfig(config *StackConfigFile) error {
-	if config == nil {
-		return errors.New("stack config cannot be nil")
-	}
-
-	// Check if we have any units or stacks
-	if len(config.Units) == 0 && len(config.Stacks) == 0 {
-		return errors.New("stack config must contain at least one unit or stack")
-	}
-
-	validationErrors := &errors.MultiError{}
-	if err := validateUnits(config.Units); err != nil {
-		validationErrors.Append(err)
-	}
-	if err := validateStacks(config.Stacks); err != nil {
-		validationErrors.Append(err)
-	}
-	return validationErrors.ErrorOrNil()
-}
-
-// validateUnits validates all units in the configuration
-func validateUnits(units []*Unit) error {
-	if len(units) == 0 {
-		return nil
-	}
-	validationErrors := &errors.MultiError{}
-	// Pre-allocate maps with known capacity to avoid resizing
-	names := make(map[string]bool, len(units))
-	paths := make(map[string]bool, len(units))
-
-	for i, unit := range units {
-		if unit == nil {
-			validationErrors.Append(errors.Errorf("unit at index %d is nil", i))
-			continue
-		}
-
-		name := strings.TrimSpace(unit.Name)
-		path := strings.TrimSpace(unit.Path)
-		source := strings.TrimSpace(unit.Source)
-
-		// Validate name
-		if name == "" {
-			validationErrors.Append(errors.Errorf("unit at index %d has empty name", i))
-		}
-
-		// Validate source
-		if source == "" {
-			validationErrors.Append(errors.Errorf("unit '%s' has empty source", unit.Name))
-		}
-
-		// Validate path
-		if path == "" {
-			validationErrors.Append(errors.Errorf("unit '%s' has empty path", unit.Name))
-		}
-
-		// Check for duplicates
-		if names[name] {
-			validationErrors.Append(errors.Errorf("duplicate unit name found: '%s'", unit.Name))
-		}
-
-		if paths[path] {
-			validationErrors.Append(errors.Errorf("duplicate unit path found: '%s'", unit.Path))
-		}
-
-		// Save non-empty values for uniqueness check
-		if name != "" {
-			names[name] = true
-		}
-
-		if path != "" {
-			paths[path] = true
-		}
-	}
-	return validationErrors.ErrorOrNil()
-}
-
-// validateStacks validates all stacks in the configuration
-func validateStacks(stacks []*Stack) error {
-	if len(stacks) == 0 {
-		return nil
-	}
-	validationErrors := &errors.MultiError{}
-	// Pre-allocate maps with known capacity to avoid resizing
-	names := make(map[string]bool, len(stacks))
-	paths := make(map[string]bool, len(stacks))
-
-	for i, stack := range stacks {
-		if stack == nil {
-			validationErrors.Append(errors.Errorf("stack at index %d is nil", i))
-			continue
-		}
-
-		name := strings.TrimSpace(stack.Name)
-		path := strings.TrimSpace(stack.Path)
-		source := strings.TrimSpace(stack.Source)
-
-		// Validate name
-		if name == "" {
-			validationErrors.Append(errors.Errorf("stack at index %d has empty name", i))
-		}
-
-		// Validate source
-		if source == "" {
-			validationErrors.Append(errors.Errorf("stack '%s' has empty source", stack.Name))
-		}
-
-		// Validate path
-		if path == "" {
-			validationErrors.Append(errors.Errorf("stack '%s' has empty path", stack.Name))
-		}
-
-		// Check for duplicates
-		if names[name] {
-			validationErrors.Append(errors.Errorf("duplicate stack name found: '%s'", stack.Name))
-		}
-
-		if paths[path] {
-			validationErrors.Append(errors.Errorf("duplicate stack path found: '%s'", stack.Path))
-		}
-
-		// Save non-empty values for uniqueness check
-		if name != "" {
-			names[name] = true
-		}
-
-		if path != "" {
-			paths[path] = true
-		}
-	}
-	return validationErrors.ErrorOrNil()
-}
-
+// processLocals processes the locals block in the stack file.
 func processLocals(parser *ParsingContext, opts *options.TerragruntOptions, file *hclparse.File) error {
 	localsBlock, err := file.Blocks(MetadataLocals, false)
 
@@ -378,4 +461,36 @@ func processLocals(parser *ParsingContext, opts *options.TerragruntOptions, file
 	parser.Locals = &localsAsCtyVal
 
 	return nil
+}
+
+// listStackFiles searches for stack files in the specified directory.
+//
+// The function walks through the given directory to find files that match the
+// default stack file name. It optionally follows symbolic links based on the
+// provided Terragrunt options.
+func listStackFiles(opts *options.TerragruntOptions, dir string) ([]string, error) {
+	walkWithSymlinks := opts.Experiments.Evaluate(experiment.Symlinks)
+	walkFunc := filepath.Walk
+	if walkWithSymlinks {
+		walkFunc = util.WalkWithSymlinks
+	}
+
+	opts.Logger.Debugf("Searching for stack files in %s", dir)
+	var stackFiles []string
+	// find all defaultStackFile files
+	if err := walkFunc(dir, func(path string, info os.FileInfo, err error) error {
+
+		if info.IsDir() {
+			return nil
+		}
+		if strings.HasSuffix(path, defaultStackFile) {
+			opts.Logger.Debugf("Found stack file %s", path)
+			stackFiles = append(stackFiles, path)
+		}
+		return nil
+	}); err != nil {
+		return nil, err
+	}
+
+	return stackFiles, nil
 }
