+++ config/stack_validation_test.go
